# Conflicts:
#	gorgone/gorgone/modules/centreon/mbi/libs/bi/BIHostStateEvents.pm
#	gorgone/gorgone/modules/centreon/mbi/libs/bi/BIServiceStateEvents.pm

+++ broker/core/inc/com/centreon/broker/config/state.hh
@@ -48,7 +48,7 @@ class state {
   std::string _module_dir;
   std::list<std::string> _module_list;
   std::map<std::string, std::string> _params;
-  int _poller_id;
+  uint64_t _poller_id;
   std::string _poller_name;
   size_t _pool_size;
 
@@ -132,8 +132,8 @@ class state {
   std::list<std::string> const& module_list() const noexcept;
   std::map<std::string, std::string>& params() noexcept;
   const std::map<std::string, std::string>& params() const noexcept;
-  void poller_id(int id) noexcept;
-  int poller_id() const noexcept;
+  void poller_id(uint64_t id) noexcept;
+  uint64_t poller_id() const noexcept;
   void pool_size(int size) noexcept;
   int pool_size() const noexcept;
   void poller_name(const std::string& name);
